chore: revert to sharing tip prop for string or snippet
Uses the same method as here:
https://github.com/sveltejs/svelte/issues/9774#issuecomment-2266343301
to tell if the property is a string or Snippet, which is both simpler and
avoids extra migration for clients.
