+++ packages/renderer/src/lib/preferences/PreferencesKubernetesContextsRendering.svelte
@@ -320,7 +320,7 @@ async function connect(contextName: string): Promise<void> {
                       <div class="ml-1 text-xs text-[var(--pd-status-dead)]" aria-label="Error">
                         <Tooltip>
                           <div>ERROR</div>
-                          {#snippet tipSnippet()}
+                          {#snippet tip()}
                             <div class="p-2">
                               {#if context.errorMessage}
                                 {#each context.errorMessage.split('\n').filter(l => l) as line, index (index)}
